feat: 'ToastUI 적용'
- ToastUI 적용
- 이미지 폴더 public으로 이동
- 기존 폴더 example 폴더로 이동
