--- src/app/(main)/(article)/blogEditor/InsertToolbar.css
@@ -1,29 +1,0 @@
-.insert-toolbar {
-    display: flex;
-    justify-content: space-between;
-    align-items: flex-start;
-    padding: 12px;
-    border-bottom: 1px solid #ddd;
-    background-color: #f8f8f8;
-}
-
-.toolbar-left,
-.toolbar-right {
-    display: flex;
-    gap: 8px;
-    flex-wrap: wrap;
-}
-
-.toolbar-right button {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    font-size: 12px;
-    color: #333;
-}
-
-.toolbar-right img {
-    width: 32px;
-    height: 32px;
-    margin-bottom: 4px;
-}
