+++ src/app/(main)/(article)/example_byebye/blogEditor_byebye/InsertToolbar.css
@@ -1,0 +1,36 @@
+.insert-toolbar {
+    display: flex;
+    justify-content: space-between;
+    align-items: flex-start;
+    padding: 12px;
+    border-bottom: 1px solid #ddd;
+    background-color: #fff;
+}
+
+.toolbar-left,
+.toolbar-right {
+    display: flex;
+    gap: 8px;
+    align-items: center;
+    /*flex-wrap: wrap;*/
+}
+
+.toolbar-left button {
+    all: unset; /* 기본 스타일 제거 */
+    cursor: pointer;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: 0;
+    margin: 0;
+}
+
+.toolbar-right button {
+    all: unset; /* 기본 스타일 제거 */
+    cursor: pointer;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: 0;
+    margin: 0;
+}
