+++ src/app/(main)/(article)/example_byebye/blogEditor_byebye/TopBar.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import './TopBar.css';
+
+const TopBar = ({ onPublish }) => (
+  <div className="top-bar">
+    <div className="top-bar__logo">
+      <img src="/image/article/logo/nblog.png" alt="N blog" />
+      <span></span>
+    </div>
+    <button className="top-bar__publish" onClick={onPublish}>
+      발행
+    </button>
+  </div>
+);
+
+export default TopBar;
